requirements.txt
gradio==5.13.1
browser-use==0.1.18
langchain-google-genai==2.0.8
python-dotenv==1.0.1